+++ Parking_lot_management-system-uge-10-11/Controllers/Parking_Lot_structursController.cs
@@ -23,6 +23,13 @@ namespace Parking_lot_management_system_uge_10_11.Controllers
         [Authorize]
         public IActionResult GetAllParking_Lot_Structur()
         {
+            var UserTypeID = User.FindFirst("UserTypeID")?.Value;
+
+            if (1 != int.Parse(UserTypeID))
+            {
+                return StatusCode(403, "Permission denied");
+            }
+
             var parking_Lot_Structur = parking_Lot_StructursRepository.GetAllparking_Lot_Structur();
 
             if (!ModelState.IsValid)
@@ -41,6 +48,12 @@ namespace Parking_lot_management_system_uge_10_11.Controllers
         [Authorize]
         public IActionResult GetParking_Lot_StructurbyOrganisation(int OrganisationID)
         {
+            var OrganisationId = User.FindFirst("OrganisationId")?.Value;
+
+            if (OrganisationID != int.Parse(OrganisationId))
+            {
+                return StatusCode(403, "Permission denied");
+            }
 
             var parking_Lot_Structur = parking_Lot_StructursRepository.Getparking_Lot_StructurByOrganisationId(OrganisationID);
 
@@ -54,12 +67,44 @@ namespace Parking_lot_management_system_uge_10_11.Controllers
             }
         }
 
+        [HttpGet("/parking_Lot_Structur/partingid{partingid}")]
+        [ProducesResponseType(200, Type = typeof(IEnumerable<Parking_Lot_Structur>))]
+        [ProducesResponseType(400)]
+        [Authorize]
+        public IActionResult GetParking_Lot_Structurbypartingid(int partingid)
+        {
+            var OrganisationId = User.FindFirst("OrganisationId")?.Value;
+
+            var parking_Lot_Structur = parking_Lot_StructursRepository.Getparking_Lot_StructurByID(partingid);
+            int parking_structurID = parking_Lot_Structur.OrganisationId;
+
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            if (parking_structurID != int.Parse(OrganisationId))
+            {
+                return StatusCode(403, "Permission denied");
+            }
+            else
+            {
+                return Ok(parking_Lot_Structur);
+            }
+        }
+
         [HttpPost("/parking_Lot_Structur/CreateParking_Lot_Structur")]
         [ProducesResponseType(200)]
         [ProducesResponseType(404)]
         [Authorize]
         public IActionResult CreateParking_Lot_Structur([FromBody] Parking_Lot_Structur parking)
         {
+            var OrganisationId = User.FindFirst("OrganisationId")?.Value;
+
+            if (parking.OrganisationId != int.Parse(OrganisationId))
+            {
+                return StatusCode(403, "Permission denied");
+            }
+
             if (parking == null)
             {
                 return BadRequest(ModelState);
@@ -97,6 +142,13 @@ namespace Parking_lot_management_system_uge_10_11.Controllers
                 ModelState.AddModelError("", "id did not exist");
                 return StatusCode(500, ModelState);
             }
+            var parking_Lot_StructurToUpdate = parking_Lot_StructursRepository.Getparking_Lot_StructurByID(parking.Parking_lot_Structur_ID);
+            var OrganisationId = User.FindFirst("OrganisationId")?.Value;
+
+            if (parking_Lot_StructurToUpdate.OrganisationId != int.Parse(OrganisationId))
+            {
+                return StatusCode(403, "Permission denied");
+            }
 
             parking_Lot_StructursRepository.Updateparking_Lot_Structur(parking);
             return Ok("parking Lot Structur Successfully Updated");
@@ -115,6 +167,13 @@ namespace Parking_lot_management_system_uge_10_11.Controllers
             }
 
             var parking_Lot_StructurToDelete = parking_Lot_StructursRepository.Getparking_Lot_StructurByID(parking_Lot_StructurID);
+
+            var OrganisationId = User.FindFirst("OrganisationId")?.Value;
+
+            if (parking_Lot_StructurToDelete.OrganisationId != int.Parse(OrganisationId))
+            {
+                return StatusCode(403, "Permission denied");
+            }
 
             if (!parking_Lot_StructursRepository.Deleteparking_Lot_Structur(parking_Lot_StructurToDelete))
             {
